[ chore ] 메뉴 등록 컴포넌트 경로 변경
